Rename INARA workspace route from ghostnet to inara
