+++ src/client/pages/inara/search.js
@@ -7,18 +7,18 @@ const navItems = [
   {
     name: 'Search',
     icon: 'search',
-    url: '/ghostnet/search',
-    active: false
+    url: '/inara/search',
+    active: true
   },
   {
     name: 'Outfitting',
     icon: 'wrench',
-    url: '/ghostnet/outfitting',
-    active: true
+    url: '/inara/outfitting',
+    active: false
   }
 ]
 
-export default function GhostnetOutfittingPage () {
+export default function GhostnetSearchPage() {
   useEffect(() => {
     if (typeof document === 'undefined' || !document.body) return undefined
     document.body.classList.add('ghostnet-theme')
@@ -31,15 +31,15 @@ export default function GhostnetOutfittingPage () {
         <div className={styles.ghostnet}>
           <div className={styles.hero}>
             <div className={styles.heroHeader}>
-              <h1 className={styles.heroTitle}>Outfitting Tools</h1>
+              <h1 className={styles.heroTitle}>Signal Search</h1>
               <p className={styles.heroSubtitle}>
-                Ship build intelligence is in fabrication. Stay tuned for modular loadouts and curated upgrade paths.
+                Global lookup is recalibrating to the new assimilation backbone. Search returns soon.
               </p>
             </div>
           </div>
 
           <div className={styles.shell}>
-            <div className={styles.placeholder}>Outfitting consoles are coming online soon.</div>
+            <div className={styles.placeholder}>General search is temporarily disabled.</div>
           </div>
         </div>
       </Panel>
